Only push/pull variables when necessary
In exceptional cases (but occuring in Brian's test suite), a variable
is only used in a run_regularly operation. In that case, it does not
need to be pushed to the device, and a StateMonitor does neither need
to pull it, nor to copy/convert between Brian2GeNN and Brian2 data
structures.